BE/#42 `AuthenticationPriciple`이 필요한 API에 사용자 식별 로직 붙이기 && `getPlaylistWithTracks()`에서 플레이리스트 소유 정보 추가 반환 (#85)
* Refactor: `GenreController` - userId 파라미터 전달방식 수정

* Fix: `AuthenticationPriciple`을 통한사용자 식별로직 추가

* Comment: `PlaylistService` 주석 추가

* Refacotr: `TagController` 파라미터 전달 방식 수정

* Fix: `TrackService` API에 사용자 식별로직 추가

* Refactor: 파라미터와 로직코드 가시적 구분을 위해 개행 추가

* Feat: 기본 경로들 whitelist에 추가

* Fix: `PlaylistService` - `getPlaylistWithTracks()` 특정 플레이리스트 정보 가져오기에서 `isOwner` 필드 추가

+++ src/main/java/com/example/dreamvalutbackend/domain/playlist/service/PlaylistService.java
@@ -50,10 +50,11 @@ public class PlaylistService {
     private final LikeRepository likeRepository;
 
     @Transactional
-    public PlaylistResponseDto createPlaylist(CreatePlaylistRequestDto createPlaylistRequestDto) {
+    public PlaylistResponseDto createPlaylist(CreatePlaylistRequestDto createPlaylistRequestDto, Long userId) {
+
         // 유저 가져오기
-        User user = userRepository.findById(1L)
-                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + 1L));
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + userId));
 
         // Playlist 생성
         Playlist playlist = Playlist.builder()
@@ -69,6 +70,7 @@ public class PlaylistService {
 
     @Transactional(readOnly = true)
     public Page<PlaylistWithTracksOverviewResponseDto> getPlaylistsWithTracksOverview(String type, Pageable pageable) {
+
         // type에 따라서 가져올 Query Function 명시
         Map<String, Function<Pageable, Page<Playlist>>> typeToQueryFunction = Map.of(
                 "curated", playlistRepository::findByIsCuratedTrue,
@@ -92,18 +94,18 @@ public class PlaylistService {
 
     @Transactional(readOnly = true)
     public PlaylistWithTracksResponseDto getPlaylistWithTracks(Long playlistId, Pageable pageable, Long userId) {
+
+        // ID로 Playlist 찾기
         Playlist playlist = playlistRepository.findById(playlistId)
                 .orElseThrow(() -> new EntityNotFoundException("Playlist not found with id: " + playlistId));
 
-        // TODO: 로그인한 유저 정보 가져오기
-        User user = userRepository.findById(1L)
-                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + 1L));
-
         // Playlist이 비공개이고 로그인한 유저와 Playlist의 유저가 다르면 예외 발생
-        if (!playlist.getIsPublic() && !playlist.getUser().equals(user)) {
+        Boolean isOwner = playlist.getUser().getUserId().equals(userId);
+        if (!playlist.getIsPublic() && !isOwner) {
             throw new SecurityException("User not authorized to view this playlist");
         }
 
+        // Playlist에 해당하는 Track들 가져오기
         Page<TrackResponseDto> tracks = playlistTrackRepository.findAllByPlaylistId(playlistId, pageable)
                 .map(playlistTrack -> {
                     // Track 가져오기
@@ -121,59 +123,63 @@ public class PlaylistService {
                     return TrackResponseDto.toDto(track, trackDetail, likes, likesFlag);
                 });
 
-        return PlaylistWithTracksResponseDto.toDto(playlist, tracks);
+
+        return PlaylistWithTracksResponseDto.toDto(playlist, tracks, isOwner);
     }
 
     @Transactional
     public PlaylistResponseDto updatePlaylistName(Long playlistId,
-            UpdatePlaylistNameRequestDto updatePlaylistNameRequestDto) {
+            UpdatePlaylistNameRequestDto updatePlaylistNameRequestDto, Long userId) {
+
+        // ID로 Playlist 찾기
         Playlist playlist = playlistRepository.findById(playlistId)
                 .orElseThrow(() -> new EntityNotFoundException("Playlist not found with id: " + playlistId));
 
-        // TODO: 로그인한 유저와 Playlist의 유저가 같은지 확인
-        // 어떤 필드로 검증 할 것인지는 JWT 토큰에 담긴 정보에 따라서 결정
-        // if (!playlist.getUser().getUserName().equals("temp")) {
-        // throw new SecurityException("User not authorized to update this playlist");
-        // }
+        // Playlist의 유저와 로그인한 유저가 다르면 예외 발생
+        User playlistOwner = playlist.getUser();
+        if (!playlistOwner.getUserId().equals(userId)) {
+            throw new SecurityException("User not authorized to update this playlist");
+        }
 
+        // Playlist 이름 업데이트
         playlist.updatePlaylistName(updatePlaylistNameRequestDto.getPlaylistName());
 
         return PlaylistResponseDto.toDto(playlist);
     }
 
     @Transactional
-    public void deletePlaylist(Long playlistId) {
+    public void deletePlaylist(Long playlistId, Long userId) {
+
         // ID로 Playlist 찾기
         Playlist playlist = playlistRepository.findById(playlistId)
                 .orElseThrow(() -> new EntityNotFoundException("Playlist not found with id: " + playlistId));
 
-        // TODO: 로그인한 유저와 Playlist의 유저가 같은지 확인
-        // 어떤 필드로 검증 할 것인지는 JWT 토큰에 담긴 정보에 따라서 결정
-        // if (!playlist.getUser().getUserName().equals("temp")) {
-        // throw new SecurityException("User not authorized to delete this playlist");
-        // }
+        // Playlist의 유저와 로그인한 유저가 다르면 예외 발생
+        User playlistOwner = playlist.getUser();
+        if (!playlistOwner.getUserId().equals(userId)) {
+            throw new SecurityException("User not authorized to delete this playlist");
+        }
 
         // 찾은 Playlist와 연관된 PlaylistTrack 삭제
         playlistTrackRepository.deleteByPlaylist(playlist);
-
-        // 찾은 Playlist와 연관된 MyPlaylist 삭제
-        myPlaylistRepository.deleteByPlaylist(playlist);
 
         // 찾은 Playlist 삭제
         playlistRepository.delete(playlist);
     }
 
     @Transactional
-    public void addTrackToPlaylist(Long playlistId, AddTrackToPlaylistRequestDto addTrackToPlaylistRequestDto) {
+    public void addTrackToPlaylist(Long playlistId,
+            AddTrackToPlaylistRequestDto addTrackToPlaylistRequestDto, Long userId) {
+
         // ID로 Playlist 찾기
         Playlist playlist = playlistRepository.findById(playlistId)
                 .orElseThrow(() -> new EntityNotFoundException("Playlist not found with id: " + playlistId));
 
-        // TODO: 로그인한 유저와 Playlist의 유저가 같은지 확인
-        // 어떤 필드로 검증 할 것인지는 JWT 토큰에 담긴 정보에 따라서 결정
-        // if (!playlist.getUser().getUserName().equals("temp")) {
-        // throw new SecurityException("User not authorized to delete this playlist");
-        // }
+        // Playlist의 유저와 로그인한 유저가 다르면 예외 발생
+        User playlistOwner = playlist.getUser();
+        if (!playlistOwner.getUserId().equals(userId)) {
+            throw new SecurityException("User not authorized to delete this playlist");
+        }
 
         // ID로 Track 찾기
         Track track = trackRepository.findById(addTrackToPlaylistRequestDto.getTrackId())
@@ -194,16 +200,17 @@ public class PlaylistService {
     }
 
     @Transactional
-    public void deleteTrackFromPlaylist(Long playlistId, Long trackId) {
+    public void deleteTrackFromPlaylist(Long playlistId, Long trackId, Long userId) {
+
         // ID로 Playlist 찾기
         Playlist playlist = playlistRepository.findById(playlistId)
                 .orElseThrow(() -> new EntityNotFoundException("Playlist not found with id: " + playlistId));
 
-        // TODO: 로그인한 유저와 Playlist의 유저가 같은지 확인
-        // 어떤 필드로 검증 할 것인지는 JWT 토큰에 담긴 정보에 따라서 결정
-        // if (!playlist.getUser().getUserName().equals("temp")) {
-        // throw new SecurityException("User not authorized to delete this playlist");
-        // }
+        // Playlist의 유저와 로그인한 유저가 다르면 예외 발생
+        User playlistOwner = playlist.getUser();
+        if (!playlistOwner.getUserId().equals(userId)) {
+            throw new SecurityException("User not authorized to delete this playlist");
+        }
 
         // ID로 Track 찾기
         Track track = trackRepository.findById(trackId)
@@ -219,10 +226,11 @@ public class PlaylistService {
     }
 
     @Transactional
-    public void followPlaylist(Long playlistId) {
-        // TODO: 로그인한 유저 정보 가져오기
-        User user = userRepository.findById(1L)
-                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + 1L));
+    public void followPlaylist(Long playlistId, Long userId) {
+
+        // ID로 User 찾기
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + userId));
 
         // ID로 Playlist 찾기
         Playlist playlist = playlistRepository.findById(playlistId)
@@ -242,10 +250,11 @@ public class PlaylistService {
     }
 
     @Transactional
-    public void unfollowPlaylist(Long playlistId) {
-        // TODO: 로그인한 유저 정보 가져오기
-        User user = userRepository.findById(1L)
-                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + 1L));
+    public void unfollowPlaylist(Long playlistId, Long userId) {
+
+        // ID로 User 찾기
+        User user = userRepository.findById(userId)
+                .orElseThrow(() -> new EntityNotFoundException("User not found with id: " + userId));
 
         // ID로 Playlist 찾기
         Playlist playlist = playlistRepository.findById(playlistId)
